expose loading_stategy in rust
