+++ keyvi/bin/keyvi_c/c_api.cpp
@@ -103,6 +103,15 @@ keyvi_dictionary* keyvi_create_dictionary(const char* filename) {
   }
 }
 
+keyvi_dictionary* keyvi_create_dictionary_with_loading_strategy(const char* filename, keyvi::dictionary::LoadingStrategy loading_strategy) {
+  try {
+    return new keyvi_dictionary(Dictionary(filename, loading_strategy));
+  } catch (const std::exception& e) {
+    std::cerr << e.what() << '\n';
+    return nullptr;
+  }
+}
+
 void keyvi_dictionary_destroy(const keyvi_dictionary* dict) {
   delete dict;
 }
