+++ rust/build.rs
@@ -53,9 +53,11 @@ fn main() {
         .enable_cxx_namespaces()
         .layout_tests(true)
         .rustified_enum("keyvi::compression::CompressionAlgorithm")
+        .rustified_enum("keyvi::dictionary::LoadingStrategy")
         .allowlist_function("keyvi_bytes_destroy")
         .allowlist_function("keyvi_string_destroy")
         .allowlist_function("keyvi_create_dictionary")
+        .allowlist_function("keyvi_create_dictionary_with_loading_strategy")
         .allowlist_function("keyvi_dictionary_destroy")
         .allowlist_function("keyvi_dictionary_get")
         .allowlist_function("keyvi_dictionary_get_all_items")
